Update main.py
Now, the cost of attribution works. The algorithm is almost complete. The only problem is that it starts reading the tasks matrix from the beginning. This is a problem because when the last tasks are going to be attributed, there are no more resources available.
I'll try to read the tasks randomly later